some code fix

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-int i = 0;
+size_t i = 0;
 while (h)
 {
 if (h->str == NULL)
@@ -19,7 +19,7 @@ printf("[0] (nil)\n");
 }
 else
 {
-printf("[%d] %s \n", h->len, h->str);
+printf("[%u] %s \n", h->len, h->str);
 }
 h = h->next;
 i++;
